zd44649: Rescue from exception instead of StandardError
There's a theory that something is raising an "Exception" instead of a
"StandardError" and that this isn't being handled properly by the Rails
app. This causes the response handling in nginx to get confused and the
user ends up seeing the wrong kind of error page.

There's also some vague stuff on the internet saying you shouldn't
rescue from Exception (because it includes things like SyntaxErrors and
SIGTERMs), but I'm can't find a canonical source for rails' rescue_from
helper. Seems to me that we should at least know what the application
will do if it sees an something that inherits Exception rather than
StandardError...

NOTE: We tried deploying this to perf to test it, and it didn't seem to
help, so ¯\_(ツ)_/¯.